Query backend with correct date in customization page
